# Conflicts:
#	domain/src/main/java/com/moscow/domain/model/Movie.kt
#	domain/src/test/java/com/moscow/domain/usecase/actor/GetActorBestMoviesUseCaseTest.kt
#	domain/src/test/java/com/moscow/domain/usecase/actor/GetActorDetailsUseCaseTest.kt
#	domain/src/test/java/com/moscow/domain/usecase/login/LoginWithUsernameAndPasswordUseCaseTest.kt
#	domain/src/test/java/com/moscow/domain/usecase/review/GetReviewsUseCaseTest.kt

+++ presentation/src/main/java/com/moscow/cineverse/screen/movieSeriesDetails/RatingBar.kt
@@ -22,7 +22,7 @@ import com.moscow.cinverse.presentation.R
 
 @Composable
 fun MovieRatingBar(
-    rating: Int,
+    rating: Float,
     onRatingChanged: (Float) -> Unit,
     modifier: Modifier = Modifier,
     starSize: Dp = 16.dp,
@@ -60,6 +60,6 @@ fun MovieRatingBar(
 @Composable
 private fun PreviewRatingBar() {
     CineVerseTheme {
-        MovieRatingBar(3, {})
+        MovieRatingBar(3.0f, {})
     }
 }
